Move FlattenElementTreeNode to its own file
